轉單修正2

+++ app/src/main/java/com/example/motoapp/GetTaskFrg.java
@@ -55,7 +55,7 @@ public class GetTaskFrg extends Activity implements GestureDetector.OnGestureLis
 	Button button_DoneList;
 	GestureDetector detector;
 	ArrayList Value;
-
+	String ObuID;
 	private final int REQUEST_CODE = 0xa1;
 	private boolean isSingleSacn = false;
 	@Override
@@ -269,42 +269,43 @@ public class GetTaskFrg extends Activity implements GestureDetector.OnGestureLis
 
 				try {
 					String status = json.getString("Result");
-
+					Log.e("轉單status",status);
 					if (status.equals("1")) {
 						//cCaseID,cOrderID,cCustAddress,cDistance,cSize,cItemCount,cRequestDate,cType
-						/*
-						String ObuID = objEdit.getText().toString();
-						// 刪除掉原本有的案件
-						dbLocations objDB;
+						ObuID = objEdit.getText().toString();
 						objDB = new dbLocations(context);
 						objDB.openDB();
 						objDB.Delete("tblTask", "cOrderID='"+ObuID+"'");
-						*/
-						// 刪除掉原本有的案件
-						objDB = new dbLocations(context);
-						objDB.openDB();
-						clsTask objT = objDB.LoadTask(json.getString("caseID"));
-						Log.e("cOrderID",objT.OrderID);
-						//objDB.Delete("tblTask", "cOrderID='"+strData+"'");
-						objDB.Delete("tblTask", "cOrderID='"+objT.OrderID+"'");
-						objDB.DBClose();
-
-						objDB = new dbLocations(context);
-						objDB.openDB();
-
 						String customer_name  = setEncryp (json.getString("customer_name"));
 						String recipient_name = setEncryp(json.getString("recipient_name")) ;
 						String recipient_phoneNo = setEncryp(json.getString("recipient_phoneNo"));
 						String recipient_address = setEncryp(json.getString("recipient_address")) ;
-						objDB.InsertTaskAllData(new Object[]{json.getString("caseID"),objEdit.getText().toString(),"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
+						objDB.InsertTaskAllData(new Object[]{json.getString("caseID"), ObuID ,"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
 						//objDB.InsertTaskAllData(new Object[]{json.getString("caseID"),contents,"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
 						objDB.DBClose();
-
-
-						Toast.makeText(GetTaskFrg.this,"取得"+objEdit.getText().toString()+"資料！",Toast.LENGTH_SHORT).show();
+						Toast.makeText(GetTaskFrg.this,"取得"+ObuID+"資料！",Toast.LENGTH_SHORT).show();
 						//clsDialog.Show(context, "提示", "取得案件資料！");
 						Log.e("Array", String.valueOf(CaptureActivity.num));
 
+						objDB = new dbLocations(context);
+						objDB.openDB();
+						clsTask objT = objDB.LoadTask(json.getString("caseID"));
+						if(!objT.equals("") && objT != null){
+							//objDB = new dbLocations(context);
+							objDB.openDB();
+							objDB.Delete("tblTask", "cOrderID='"+objT.OrderID+"'");
+							Log.e("刪除2",ObuID);
+							customer_name  = setEncryp (json.getString("customer_name"));
+							recipient_name = setEncryp(json.getString("recipient_name")) ;
+							recipient_phoneNo = setEncryp(json.getString("recipient_phoneNo"));
+							recipient_address = setEncryp(json.getString("recipient_address")) ;
+							objDB.InsertTaskAllData(new Object[]{json.getString("caseID"), ObuID ,"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
+							//objDB.InsertTaskAllData(new Object[]{json.getString("caseID"),contents,"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
+							objDB.DBClose();
+							Toast.makeText(GetTaskFrg.this,"取得"+ObuID+"資料！",Toast.LENGTH_SHORT).show();
+							//clsDialog.Show(context, "提示", "取得案件資料！");
+							Log.e("Array", String.valueOf(CaptureActivity.num));
+						}
 
 					}
 					if (status.equals("2")) {
@@ -463,13 +464,11 @@ public class GetTaskFrg extends Activity implements GestureDetector.OnGestureLis
 									//clsTask objT = objDB.LoadTask(Application.strCaseID);
 									*/
 									// 刪除掉原本有的案件
+									dbLocations objDB;
 									objDB = new dbLocations(context);
 									objDB.openDB();
-									clsTask objT = objDB.LoadTask(j.getString("caseID"));
-									Log.e("掃描caseID",j.getString("caseID"));
-									Log.e("cOrderID",objT.OrderID);
-									objDB.Delete("tblTask", "cOrderID='"+objT.OrderID+"'");
-
+									objDB.Delete("tblTask", "cOrderID='"+contents+"'");
+									//objDB.Delete("tblTask", "cOrderID='"+objT.OrderID+"'");
 									//加入轉單後的案件
 									String customer_name = setEncryp(j.getString("customer_name"));
 									String recipient_name = setEncryp(j.getString("recipient_name"));
@@ -479,10 +478,40 @@ public class GetTaskFrg extends Activity implements GestureDetector.OnGestureLis
 									objDB.InsertTaskAllData(new Object[]{j.getString("caseID"), contents, "", "", j.getString("size"), j.getString("item_count"), j.getString("status_time"), "1", customer_name, "", recipient_name, recipient_phoneNo, recipient_address, j.getString("request_time"), j.getString("pay_type_MD"), j.getString("pay_amount_MD"), j.getString("cash_on_delivery")});
 									objDB.close();
 
-									//Toast.makeText(GetTaskFrg.this,"取得"+Application.getTask+"資料！",Toast.LENGTH_SHORT).show();
+									runOnUiThread(new Runnable() {
+										@Override
+										public void run() {
+											Toast.makeText(GetTaskFrg.this,"取得"+contents+"資料！",Toast.LENGTH_SHORT).show();
+										}
+									});
 									//clsDialog.Show(context, "提示", "取得案件資料！");
 									//Log.e("Array", String.valueOf(CaptureActivity.num));
 
+									objDB = new dbLocations(context);
+									objDB.openDB();
+									clsTask objT = objDB.LoadTask(j.getString("caseID"));
+									if(!objT.equals("") && objT != null){
+										//objDB = new dbLocations(context);
+										objDB.openDB();
+										objDB.Delete("tblTask", "cOrderID='"+objT.OrderID+"'");
+
+										customer_name  = setEncryp (j.getString("customer_name"));
+										recipient_name = setEncryp(j.getString("recipient_name")) ;
+										recipient_phoneNo = setEncryp(j.getString("recipient_phoneNo"));
+										recipient_address = setEncryp(j.getString("recipient_address")) ;
+										objDB.InsertTaskAllData(new Object[]{j.getString("caseID"), contents ,"","",j.getString("size"),j.getString("item_count"),j.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,j.getString("request_time"),j.getString("pay_type_MD"),j.getString("pay_amount_MD"),j.getString("cash_on_delivery")});
+										//objDB.InsertTaskAllData(new Object[]{json.getString("caseID"),contents,"","",json.getString("size"),json.getString("item_count"),json.getString("status_time"),"1",customer_name,"",recipient_name,recipient_phoneNo,recipient_address,json.getString("request_time"),json.getString("pay_type_MD"),json.getString("pay_amount_MD"),json.getString("cash_on_delivery")});
+										objDB.DBClose();
+										runOnUiThread(new Runnable() {
+											@Override
+											public void run() {
+												Toast.makeText(GetTaskFrg.this,"取得"+contents+"資料！",Toast.LENGTH_SHORT).show();
+											}
+										});
+
+										//clsDialog.Show(context, "提示", "取得案件資料！");
+										Log.e("Array", String.valueOf(CaptureActivity.num));
+									}
 
 								}
 								if (status.equals("2")) {
